Extracted collaboration extension code to new file
